refactor(step2):  패키지 구조 변경
같은성격의 클래스 묶기
