+++ src/main/java/bowling/domain/BowlingGame.java
@@ -1,5 +1,8 @@
 package bowling.domain;
 
+import bowling.domain.frame.Frame;
+import bowling.domain.frame.Frames;
+
 public class BowlingGame {
     private Frames frames;
