+++ src/main/java/bowling/domain/frame/FinalFrame.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.frame;
 
 public class FinalFrame extends Frame {
