+++ src/main/java/bowling/domain/frame/Frames.java
@@ -1,4 +1,6 @@
-package bowling.domain;
+package bowling.domain.frame;
+
+import bowling.domain.Player;
 
 import java.util.ArrayList;
 import java.util.List;
