+++ src/main/java/bowling/domain/point/Points.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.point;
 
 import java.util.ArrayList;
 import java.util.List;
