+++ src/test/java/bowling/domain/frame/FramesTest.java
@@ -1,5 +1,8 @@
-package bowling.domain;
+package bowling.domain.frame;
 
+import bowling.domain.Player;
+import bowling.domain.frame.Frame;
+import bowling.domain.frame.Frames;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
