+++ src/test/java/bowling/domain/point/PointTest.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.point;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
